+++ AxonDeepSeg/data_management/dataset_building.py
@@ -1,4 +1,4 @@
-from pathlib import Path
+
 
 from imageio import imread, imsave
 from skimage.transform import rescale