+++ AxonDeepSeg/integrity_test.py
@@ -7,9 +7,10 @@
 import json
 from pathlib import Path
 from AxonDeepSeg.testing.segmentation_scoring import *
-from time import time
 from AxonDeepSeg.apply_model import axon_segmentation
+
 from imageio import imread, imsave
+
 import AxonDeepSeg.ads_utils
 
 def integrity_test():