+++ AxonDeepSeg/testing/noise_simulation.py
@@ -1,5 +1,6 @@
 # coding: utf-8
 
+
 import numpy as np
 from skimage import io
 import os
@@ -11,8 +12,10 @@ from skimage.filters import gaussian
 from AxonDeepSeg.apply_model import axon_segmentation
 from AxonDeepSeg.testing.segmentation_scoring import *
 from imageio import imread, imsave
-import AxonDeepSeg.ads_utils
 
+from AxonDeepSeg.testing.segmentation_scoring import *
+
+import AxonDeepSeg.ads_utils
 
 def add_additive_gaussian_noise(img,mu=0,sigma=10):
 	'''