+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to the AxonDeepSeg project will be documented in this file.
 - Revert tensorflow requirekment to 1.3.0 and remove tifffile requirement
 - Remove `matplotlib.pyplot` from source code and refactor it to full OO plotting
 - Standardize path management to `pathlib` library
+- Shifted AxonDeepSeg from TensorFlow to Keras framework.
 – Add FSLeyes plugin
 
 ## [2.1] - 2018-09-25